+++ src/StateSmith.Cli/Setup/SetupUi.cs
@@ -7,11 +7,13 @@ public class SetupUi
 {
     private SetupOptions opts;
     private IAnsiConsole _console;
+    private readonly string _currentDirectory;
 
-    public SetupUi(SetupOptions opts, IAnsiConsole console)
+    public SetupUi(SetupOptions opts, IAnsiConsole console, string currentDirectory)
     {
         this.opts = opts;
         _console = console;
+        this._currentDirectory = currentDirectory;
     }
 
     public int Run()
@@ -62,7 +64,7 @@ public class SetupUi
 
     private void SetupVscodeCsx()
     {
-        new SetupVscodeCsxAction(_console, verbose: opts.Verbose).Run();
+        new SetupVscodeCsxAction(_console, verbose: opts.Verbose, currentDirectory: _currentDirectory).Run();
     }
 
     private void AskWhatToDo(string? title = null)