+++ src/StateSmith/Runner/DiscardingConsolePrinter.cs
@@ -1,8 +1,12 @@
-﻿namespace StateSmith.Runner;
+namespace StateSmith.Runner;
 
 // we may want to capture this info at some point
 public class DiscardingConsolePrinter : IConsolePrinter
 {
+    public void SetLineEnding(string lineEnding)
+    {
+    }
+
     public void WriteErrorLine(string message)
     {
     }