+++ src/StateSmith/Runner/SmRunner.cs
@@ -6,6 +6,7 @@ using StateSmith.Output.UserConfig;
 using StateSmith.Common;
 using StateSmith.SmGraph;
 using System;
+using System.Runtime.ExceptionServices;
 
 namespace StateSmith.Runner;
 
@@ -31,6 +32,8 @@ public class SmRunner : SmRunner.IExperimentalAccess
     /// figuring for regular C# projects and C# scripts (.csx).
     /// </summary>
     readonly string callerFilePath;
+
+    private ExceptionDispatchInfo? preDiagramBasedSettingsException = null;
 
     /// <summary>
     /// Constructor. Will attempt to read settings from the diagram file.
@@ -79,18 +82,19 @@ public class SmRunner : SmRunner.IExperimentalAccess
     /// </summary>
     public SmTransformer SmTransformer => diServiceProvider.GetServiceOrCreateInstance();
 
+    /// <summary>
+    /// This API is experimental and may change in the future.
+    /// </summary>
+    public ExceptionDispatchInfo? PreDiagramBasedSettingsException => preDiagramBasedSettingsException;
 
     /// <summary>
     /// Runs StateSmith. Will cause the Dependency Injection settings to be finalized.
     /// </summary>
     public void Run()
     {
-        if (settings.transpilerId == TranspilerId.NotYetSet)
-            throw new ArgumentException("TranspilerId must be set before running code generation");
-
         SmRunnerInternal.AppUseDecimalPeriod(); // done here as well to be cautious for the future
 
-        PrepareBeforeRun();
+        PrepareBeforeRun(); // finalizes dependency injection
         SmRunnerInternal smRunnerInternal = diServiceProvider.GetServiceOrCreateInstance();
         smRunnerInternal.preDiagramBasedSettingsAlreadyApplied = enablePreDiagramBasedSettings;
 
@@ -98,6 +102,11 @@ public class SmRunner : SmRunner.IExperimentalAccess
         // dispose of objects that it created.
         try
         {
+            ThrowIfPreDiagramSettingsException();
+
+            if (settings.transpilerId == TranspilerId.NotYetSet)
+                throw new ArgumentException("TranspilerId must be set before running code generation");
+
             smRunnerInternal.Run();
         }
         finally
@@ -106,7 +115,25 @@ public class SmRunner : SmRunner.IExperimentalAccess
         }
 
         if (smRunnerInternal.exception != null)
-            throw new System.Exception("Finished with failure");
+            throw new FinishedWithFailureException();
+    }
+
+    /// <summary>
+    /// Experimental API. May change in the future.
+    /// </summary>
+    public void ThrowIfPreDiagramSettingsException()
+    {
+        if (PreDiagramBasedSettingsException != null)
+        {
+            // We use SmRunnerInternal to print the exception so that it is consistent with the rest of the code.
+            SmRunnerInternal smRunnerInternal = diServiceProvider.GetServiceOrCreateInstance();
+            smRunnerInternal.PrintException(PreDiagramBasedSettingsException.SourceException);
+            if (settings.propagateExceptions)
+            {
+                PreDiagramBasedSettingsException.Throw();
+            }
+            throw new FinishedWithFailureException();
+        }
     }
 
     // ------------ private methods ----------------
@@ -122,8 +149,22 @@ public class SmRunner : SmRunner.IExperimentalAccess
         // we disable early diagram settings reading for the simulator and some tests
         if (enablePreDiagramBasedSettings)
         {
-            PreDiagramSettingsReader preDiagramSettingsReader = new(renderConfigAllVars, settings, iRenderConfig);
-            preDiagramSettingsReader.Process();
+            // Why do we do this before DiServiceProvider is set up? It is a pain to not have DI set up.
+            // A number of reasons. We need to read the settings before we can set up the DI provider.
+            // Also (less importantly), we want to read settings from the diagram so that the user can
+            // override them in a .csx file (if they choose) before running the code generator.
+            // https://github.com/StateSmith/StateSmith/issues/349
+            try
+            {
+                // Note that this may throw if the diagram is invalid.
+                PreDiagramSettingsReader preDiagramSettingsReader = new(renderConfigAllVars, settings, iRenderConfig);
+                preDiagramSettingsReader.Process();
+            }
+            catch (Exception e)
+            {
+                // NOTE! we can't print or log this exception here because dependency injection is not yet set up
+                preDiagramBasedSettingsException = ExceptionDispatchInfo.Capture(e);
+            }
         }
 
         AlgoOrTranspilerUpdated();
@@ -174,7 +215,10 @@ public class SmRunner : SmRunner.IExperimentalAccess
         new AlgoTranspilerCustomizer().Customize(diServiceProvider, settings.algorithmId, settings.transpilerId, settings.algoBalanced1);
     }
 
-    // exists just for testing. can be removed in the future.
+    /// <summary>
+    /// Finalizes dependency injection settings.
+    /// exists just for testing. can be removed in the future.
+    /// </summary>
     internal void PrepareBeforeRun()
     {
         diServiceProvider.BuildIfNeeded();